+++ roadmap.txt
@@ -1,4 +1,4 @@
-singe time jobs (max runs)
+singe time jobs (max runs) ok
 preventOverrun
 error handling, add a onError option
 docs comments
